organize components folder
