+++ platform/src/common/components/Collocation/AddMonitor/Calendar/index.jsx
