+++ platform/src/common/components/Collocation/AddMonitor/Skeletion/Calendar.jsx
