+++ platform/src/common/components/Collocation/AddMonitor/Skeletion/Table.jsx
