+++ platform/src/common/components/Collocation/AddMonitor/Table/DataTable.jsx
