+++ platform/src/common/components/Collocation/AddMonitor/Table/Pagination.jsx
