+++ platform/src/common/components/Collocation/AddMonitor/Table/SearchBar.jsx
